refactor(vim/vimrc): change mappings for horizontal and vertical window splits
The previous `<Leader>h` for a horizontal window split had a conflict
with the default mappings of vim-gitgutter, which also prepend
`<Leader>h`. To resolve this conflict, the mapping for horizontal window
split has been changed to `<Leader>-` and accordingly, the mapping for
vertical window split has been changed to `<Leader>\`. These mappings
using the `-` and `\` keys are similar to my window split keymappings
for tmux.