fix(vim/vimrc): delete `suan/vim-instant-markdown` plugin
This plugin is nice but caused me an error when switching to a buffer
from a buffer editing a markdown file. This seems to happen when a
related npm module is not installed. I did not utilise this plugin
anyway so this has been deleted. For a record, I keep a Vundle's
installation command oneliner commented out for a while.