fix(bash/bashrc): insert `pip` aliases below pyenv's initialisation scripts
This commit fix broken Bash aliases for the `pip` command, caused by
a Homebrew update introduced in 1.3.0

As of Homebrew 1.3.0, the names of the Python 2 binaries have been
changed from `python` to `python2` and from `pip` to `pip2`. This change
differentiates between the default Python shipped with macOS and the
version installed with Homebrew.

See the following link for the details:
https://brew.sh/2017/07/31/homebrew-1.3.0/

In order to solve this, the declarations of `pip` aliases have been
inserted below the initialisations of `pyenv`.