fix(git/gitconfig): relocate `[include]` declaration location
The settings in the `~/.gitconfig.local` file should come below the
default settings in the global `~/.gitconfig` file in order for it to
become effective. The common use case for this file, which is at least
currently relevant to me, is to override the `[user]` and `[github]`
settings. For now, the `[include]` setting file is located right below
the `[user]` and `[github]` setting field.